refactor: extract SessionsConfig
