+++ crates/config/Cargo.toml
@@ -12,7 +12,7 @@ workspace = true
 
 [dependencies]
 # reth
-reth-network.workspace = true
+reth-network-types.workspace = true
 reth-prune-types.workspace = true
 
 # serde
