+++ crates/net/network-types/src/session/mod.rs
@@ -1,0 +1,4 @@
+//! Peer sessions configuration.
+
+pub mod config;
+pub use config::{SessionLimits, SessionsConfig};
